backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.1
GitPython==3.1.40
PyGithub==2.1.1
reportlab==4.0.7
Pillow==10.1.0
